working

+++ src/main/java/com/test/bu/controller/WalletController.java
@@ -1,18 +1,26 @@
 package com.test.bu.controller;
 
 
+import com.test.bu.entity.Users;
 import com.test.bu.entity.Wallet;
+import com.test.bu.service.UsersService;
 import com.test.bu.service.WalletService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
 
+import java.security.Principal;
+import java.util.List;
+
 @Controller
 @RequestMapping("/")
 public class WalletController {
 
+
    //protected static long countWalletsNumber=100000000;
+@Autowired
+private UsersService usersService;
 
     @Autowired
     private WalletService walletService;
@@ -31,12 +39,26 @@ public class WalletController {
     }
 
     @PostMapping("/newWallet")
-    public String createWallet(@ModelAttribute Wallet wallet) {
+    public String createWallet(@ModelAttribute Wallet wallet, Principal principal) {
         wallet.setNumber(walletService.getWalletsInDB() + 100_000_000);
-        /*for (:
-             ) {
-
-        }*/
+        /*List<Wallet> allWalletNumbers = walletService.getAllWalletNumbers();
+        for (Wallet n :
+                allWalletNumbers) {
+            System.out.println(n);
+        }
+        for (int i = 0; i < allWalletNumbers.size(); i++) {
+            Wallet j = allWalletNumbers.get(i);
+            long k = j.getNumber();
+            if (wallet.getNumber() == k) {
+                wallet.setNumber(wallet.getNumber()+1);
+                i--;
+            }
+        }
+        */
+        Users user = usersService.getUserByName(principal.getName());
+        user.getWalletsList().add(wallet);
+        wallet.setUserId(user);
+        usersService.update(user);
         walletService.save(wallet);
         return "redirect:wallets";
     }
@@ -55,8 +77,15 @@ public class WalletController {
     }
 
     @GetMapping("wallets/delete/{number}")
-    public String deleteWallet(@PathVariable("number") int number) {
-        walletService.delete(number);
+    public String deleteWallet(@PathVariable("number") int number, Principal principal) {
+        Users user = usersService.getUserByName(principal.getName());
+        Wallet wallet = walletService.getByNumber(number);
+        Users userFromPrincipal = wallet.getUserId();
+        if (user.getId() == userFromPrincipal.getId()) {
+            wallet.setUserId(null);
+            walletService.update(wallet);
+            walletService.delete(number);
+        }
         return "redirect:/wallets";
     }
